remove duplicate of ensureRightNetwork function
